SAT/ai-model/requirements.txt
pandas
numpy
pymongo
python-dotenv
scikit-learn
flask
flask-cors
huggingface_hub
neuralforecast
neuralprophet
gluonts
scipy
torch
transformers
accelerate
bitsandbytes
qdrant-client
sentence-transformers
groq
